+++ mne_bids/mne_bids.py
@@ -36,7 +36,7 @@ from .utils import (make_bids_basename, make_bids_folders,
 from .copyfiles import (copyfile_brainvision, copyfile_eeglab, copyfile_ctf,
                         copyfile_bti)
 
-from .io import _parse_ext, reader
+from .read import _parse_ext, reader
 from .tsv_handler import _from_tsv, _combine, _drop, _contains_row
 
 from .config import (ALLOWED_KINDS, ORIENTATION, UNITS, MANUFACTURERS,
